+++ src/audio/audio_mikmod.h
@@ -1,6 +1,6 @@
 /* MegaZeux
  *
- * Copyright (C) 2004 Gilead Kutnick <exophase@adelphia.net>
+ * Copyright (C) 2007 Alistair John Strachan <alistair@devzero.co.uk>
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -19,15 +19,15 @@
 
 /* Declarations */
 
-#ifndef __AUDIO_MODPLUG_H
-#define __AUDIO_MODPLUG_H
+#ifndef __AUDIO_MIKMOD_H
+#define __AUDIO_MIKMOD_H
 
-#include "compat.h"
+#include "../compat.h"
 
 __M_BEGIN_DECLS
 
-void init_modplug(struct config_info *conf);
+void init_mikmod(struct config_info *conf);
 
 __M_END_DECLS
 
-#endif  // __AUDIO_MODPLUG_H
+#endif  // __AUDIO_MIKMOD_H
