+++ src/audio/audio_modplug.h
@@ -1,8 +1,6 @@
 /* MegaZeux
  *
  * Copyright (C) 2004 Gilead Kutnick <exophase@adelphia.net>
- * Copyright (C) 2007 Alistair John Strachan <alistair@devzero.co.uk>
- * Copyright (C) 2018 Alice Rowan <petrifiedrowan@gmail.com>
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -19,18 +17,17 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#ifndef __AUDIO_STREAM_PCS_H
-#define __AUDIO_STREAM_PCS_H
+/* Declarations */
 
-#include "compat.h"
+#ifndef __AUDIO_MODPLUG_H
+#define __AUDIO_MODPLUG_H
+
+#include "../compat.h"
 
 __M_BEGIN_DECLS
 
-void sound(int frequency, int duration);
-void nosound(int duration);
-
-void init_pc_speaker(struct config_info *conf);
+void init_modplug(struct config_info *conf);
 
 __M_END_DECLS
 
-#endif /* __AUDIO_STREAM_PCS_H */
+#endif  // __AUDIO_MODPLUG_H
