+++ src/audio/audio_xmp.h
@@ -19,15 +19,15 @@
 
 /* Declarations */
 
-#ifndef __AUDIO_MIKMOD_H
-#define __AUDIO_MIKMOD_H
+#ifndef __AUDIO_XMP_H
+#define __AUDIO_XMP_H
 
-#include "compat.h"
+#include "../compat.h"
 
 __M_BEGIN_DECLS
 
-void init_mikmod(struct config_info *conf);
+void init_xmp(struct config_info *conf);
 
 __M_END_DECLS
 
-#endif  // __AUDIO_MIKMOD_H
+#endif  // __AUDIO_XMP_H
