+++ src/game.c
@@ -31,29 +31,30 @@
 #include <unistd.h>
 #endif
 
-#include "platform.h"
-#include "event.h"
-#include "helpsys.h"
-#include "scrdisp.h"
-#include "error.h"
-#include "idarray.h"
-#include "audio.h"
-#include "sfx.h"
-#include "idput.h"
-#include "data.h"
-#include "const.h"
-#include "game.h"
-#include "window.h"
-#include "graphics.h"
-#include "counter.h"
-#include "game2.h"
-#include "sprite.h"
-#include "world.h"
 #include "board.h"
-#include "robot.h"
-#include "fsafeopen.h"
+#include "const.h"
+#include "counter.h"
+#include "data.h"
+#include "error.h"
+#include "event.h"
 #include "extmem.h"
+#include "fsafeopen.h"
+#include "game.h"
+#include "game2.h"
+#include "graphics.h"
+#include "helpsys.h"
+#include "idarray.h"
+#include "idput.h"
+#include "platform.h"
+#include "robot.h"
+#include "scrdisp.h"
+#include "sprite.h"
 #include "util.h"
+#include "window.h"
+#include "world.h"
+
+#include "audio/audio.h"
+#include "audio/sfx.h"
 
 #define MESG_TIMEOUT 160
